third_party/WebKit/LayoutTests/inspector/profiler/cpu-profiler-native-nodes-filter-expected.txt

Tests filtering of native nodes.

(root)(1): 100000
  holder(10): 11110
    Aaaaaa(5): 20
    Bbb(6): 200
  C(7): 22000
    Ccc(9): 1000
1, 10, 10, 10, 5, 6, 7, 7, 9, 10
